Refactor: uses HOC for extension instead
