+++ packages/grafana-ui/src/components/Table/withTableStyles.tsx
@@ -1,0 +1,14 @@
+import { CellComponent, TableCellProps } from './types';
+import { TableStyles } from './styles';
+
+export const withTableStyles = (
+  CellComponent: CellComponent,
+  getExtendedStyles: (props: TableCellProps) => TableStyles
+): CellComponent => {
+  function WithTableStyles(props: TableCellProps) {
+    return CellComponent({ ...props, tableStyles: getExtendedStyles(props) });
+  }
+
+  WithTableStyles.displayName = CellComponent.displayName || CellComponent.name;
+  return WithTableStyles;
+};
